Enemy LoS checks and Projectile Min Velocity
Added checks to see if enemies are in line of sight of the player before firing - eventually can have them roaming to try and find the player if they're not in LoS. Needed to add the enemy prefab to the ignore raycast layer for this to work properly. Added checks for projectiles for minimum velocity so we can destroy them if they aren't going fast - they get destroyed if they're going below a certain speed for a specified time in script that can be changed in editor.