Splash Screen
Added a splash screen when the game loads up as the initial scene in the build settings. Adjusted the build settings so that that fades into the main menu and clicking play now loads the corridor level. Added a function to the level loader to just load next level so we don't have to get the build index from different scripts every time if we just want the next level.